+++ dbt/include/databricks/macros/materializations/incremental/incremental.sql
@@ -17,13 +17,6 @@
   {%- set target_relation = this -%}
   {%- set existing_relation = adapter.get_relation(database=this.database, schema=this.schema, identifier=this.identifier, needs_information=True) -%}
 
-  {#-- Set Overwrite Mode - does not work for warehouses --#}
-  {%- if incremental_strategy == 'insert_overwrite' and partition_by -%}
-    {%- call statement() -%}
-      set spark.sql.sources.partitionOverwriteMode = DYNAMIC
-    {%- endcall -%}
-  {%- endif -%}
-
   {#-- Run pre-hooks --#}
   {{ run_hooks(pre_hooks) }}