+++ dbt/include/databricks/macros/relations/file_format.sql
@@ -1,6 +1,4 @@
-{% macro databricks__file_format_clause() %}
+{% macro file_format_clause() %}
   {%- set file_format = config.get('file_format', default='delta') -%}
-  {%- if file_format is not none %}
-    using {{ file_format }}
-  {%- endif %}
+  using {{ file_format }}
 {%- endmacro -%}