+++ tests/integration/liquid_clustering/models/liquid_clustered.sql
@@ -1,2 +1,0 @@
-{{ config(materialized='incremental', liquid_clustered_by='id') }}
-select 1 as id, 'Joe' as name