+++ tests/integration/materialized_view/models/mv.sql
@@ -1,5 +1,0 @@
-{{ config(
-    materialized = 'materialized_view',
-) }}
-
-select id as id, count(*) as cnt from {{ ref('base') }} group by id