+++ tests/integration/persist_constraints/models/incremental_model.sql
@@ -1,7 +1,0 @@
-{{config(materialized='incremental')}}
-
-select * from {{ ref('seed') }}
-
-{% if is_incremental() %}
-  where date > (select max(date) from {{ this }})
-{% endif %}